separate out init function
