+++ src/lib/stores.ts
@@ -4,5 +4,4 @@ import type { FlightData, FlightLog } from '$lib/structs';
 export const InfoVisible = writable(true);
 export const FlightRecord = writable({ initialized: false } as FlightData);
 export const LogArray = writable([] as FlightLog[]);
-export const Init = writable(false);
 export const FlagInvalid = writable(false);
